add big patch: support of files with more than 16.7 Mio samples; multiple sampling rates of event table etc. (was biosig4c++/patches/patch_sigviewer_0.5.2.diff)

+++ sigviewer/src/base/math_utils.cpp
@@ -55,7 +55,7 @@ float64 round125(float64 value)
 }
 
 //-----------------------------------------------------------------------------
-int sampleRateToDecimalPrecision (float sample_rate)
+int sampleRateToDecimalPrecision (float64 sample_rate)
 {
     int precision = 0;
     for (; sample_rate > 10; sample_rate /= 10)
